refactor: Improve file naming conventions for SQL tools and APIs
- Rename sql-tool.ts to sql-query-handler.ts for clarity
- Rename sql-execute-tool.ts to sql-confirm-execute-tool.ts to indicate confirmation flow
- Rename api/sql-execute.ts to api/sql-confirm-execute.ts for consistency
- Update all import statements across the codebase
- Rename corresponding test files to match new naming

This refactoring makes the purpose of each module clearer:
- sql-query-handler: Handles direct SQL query execution
- sql-confirm-execute-tool: Manages token-based confirmed execution
- sql-confirm-execute API: Provides the confirmation execution endpoint

Resolves #9
