+++ src/mastra/api/sql-confirm-execute.ts
