+++ src/mastra/index.ts
@@ -1,7 +1,7 @@
 import { Mastra } from '@mastra/core/mastra';
 import { PinoLogger } from '@mastra/loggers';
 import { sqlAgent } from './agents/sql-agent';
-import { sqlExecuteRoute } from './api/sql-execute';
+import { sqlExecuteRoute } from './api/sql-confirm-execute';
 import { serverConfig } from '../config/env';
 
 export const mastra = new Mastra({
